fixed test

+++ Calculate.Tests/CalculatorTests.cs
@@ -42,7 +42,7 @@ public class CalculatorTests
     [InlineData(0.5,1,2)]
     [InlineData(2,4,2)]
     [InlineData(2,8,4)]
-    public void Divide_Success(int expected, int value1, int value2)
+    public void Divide_Success(double expected, int value1, int value2)
     {
         Assert.Equal<double>(expected, Calculator.Divide( value1, value2));
     }
